WGDev can now write to its config file
WGDev along with the 'config' command can now write to the wgdevcfg
file.  This involves a change in the calling syntax of WGDev::wgd_config
and WGDev::my_config.  They now accept as their first parameter an array
ref of keys to traverse or the keys joined together with periods.  It
also will intelligently handle numeric keys for array lookups.  The
second parameter to either function is the value to set.  Changes to
config parameters are not written to the file automatically.