Change WGDev::Command::Base's command parsing API.
The methods option_config and option_parse_config in WGDev::Command::Base
have been renamed config_options and config_parse_options respectively.
This is an API break, but fixes the conflict between the handling
of an option named 'config' and the methods for configuring parsing.
This, along with a change to WGDev::Command, fixes the handling of
the Reset command's config reset option.